move user details to constants.js
