+++ frontend/common/constants.js
@@ -1,3 +1,5 @@
+import { escapeHTML } from './utils.js';
+
 export const Links = {
     serverBaseURL:
         'http://spiderman-trivia-api.eu-west-1.elasticbeanstalk.com/api',
@@ -15,3 +17,16 @@ export const Icons = {
 export const Images = {
     loading: 'assets/images/loading.gif',
 };
+
+const rawUsername = localStorage.getItem('username') || "there";
+const username = escapeHTML(rawUsername);
+const authToken = localStorage.getItem('authToken');
+const signedIn = localStorage.getItem('signedIn') || "false";
+const profilePictureUrl = username === 'there' ? `assets/images/default-avatar.jpeg` : `https://github.com/${username}.png`;
+
+export const User = {
+    username,
+    authToken,
+    signedIn,
+    profilePictureUrl
+}
