+++ frontend/common/utils.js
@@ -1,0 +1,15 @@
+export const escapeHTML = (str)=> {
+    if(typeof(str) == 'string'){
+        return str.replace(/[&<>"']/g, function (match) {
+            return {
+                '&': '&amp;',
+                '<': '&lt;',
+                '>': '&gt;',
+                '"': '&quot;',
+                "'": '&#39;'
+            }[match];
+        });
+    }
+return "there";
+
+}
